KP-9134 Fix `personInfo` elements being found outside the current record
Despite `sickle.ListReords` iterating over the individual
`record` elements, they still have the `ListRecords` element as parent
and other records as siblings. This means that using `//` in the XPath
expressions resulted in matches being found outside the current record.
Now the search is started from the root of the current record instead.