KP-9134 Indent the modifications made to the record
Without reindenting after inserting the new elements, the modified
elements would either contain non-standard whitespace or the string
literal from which they are generated would need to be indented in a way
that does not match the surrounding code. Fixing the indentation allows
us to keep both the code and the produced diffs neat.