KP-9134 Fix organization info missing on repeat instances of person
If the same lxml Element was added as a child element to more than one
element (e.g. both licensor and IPR holder requiring the same
affiliation to be added), the element only appeared in one of the
expected places due to the fact that the same element cannot have more
than one parent. Now a new element is created every time an affiliation
is added, meaning that we no longer get affiliations without
organization info.