KP-9134 Always set records to published when uploading
Based on record https://clarino.uib.no/comedi/editor/lb-201403266 not
being published at the time of writing but still being present in the
OAI-PMH results, I assumed that Comedi's OAI-PMH API systematically
lists unpublished records. It doesn't though: even that record is listed
as published in the "show records" view, and based on everything visible
to the user, it is indeed published. Might be that the record has been
edited to that it isn't valid and thus automatically set to in-progress,
but this change does not affect all views/APIs. This is not something we
can fix in our end though, so it seems that best we can do is to try to
publish all updated records.