KP-9803 Add organizations whose full match is found
We can add organizations for which there already exists an
organizationInfo element with the same organization name (either in
Finnish or English), but will skip others.

All records that already have resourceCreationInfo element are skipped,
as we don't want to mess up manually created data.

Persons are not handled at all yet.